requirements.txt
paho-mqtt
pika
pandas